+++ cytosnake/tests/functional/test_cli.py
@@ -21,8 +21,8 @@ provided with valid user parameters. The primary goal is to ensure that the CLI 
 successfully without errors and produces the correct output.
 
 Negative Cases:
-These test cases simulate scenarios where invalid user parameters are provided. 
-The objective is to confirm that the CLI can identify and handle errors appropriately, 
+These test cases simulate scenarios where invalid user parameters are provided.
+The objective is to confirm that the CLI can identify and handle errors appropriately,
 providing informative error messages to the user.
 """
 
@@ -34,6 +34,7 @@ import shutil
 import subprocess
 
 from cytosnake.common import errors
+from cytosnake.tests.test_utils import get_raised_error
 
 
 # --------------
@@ -150,32 +151,6 @@ def get_test_data_folder(test_data_name: str) -> DataFiles:
     data_files = DataFiles(sel_test_data)
 
     return data_files
-
-
-def get_raised_error(traceback: str) -> str:
-    """Parses traceback and attempts to obtain raised exception error.
-
-    Traceback is parsed in this order:
-    1. split by new lines
-    2. grab the last line as it contains the raised exception and message
-    3. split by ":" to separate exception name and exception message
-    4. grab the first element since it contains that path to exception
-    5. split by "." and grab last element, which is the exception name
-
-    Parameters
-    ----------
-    traceback : str
-        complete traceback generated by executing CLI
-
-    Returns
-    -------
-    str
-        return raised exception error
-    """
-
-    # returns exception name, refer to function documentation to understand
-    # the order of parsing the traceback to obtain exception name.
-    return traceback.splitlines()[-1].split(":")[0].split(".")[-1]
 
 
 def prepare_dataset(
@@ -323,14 +298,14 @@ def test_one_plate_one_platemap(testing_dir) -> None:
 def test_multiplates_with_multi_platemaps(testing_dir):
     """Positive case test: Expects a 0 exit code
 
-    This test checks if the CLI can handle multiple plates and multiple plate maps 
+    This test checks if the CLI can handle multiple plates and multiple plate maps
     with a barcode as inputs.
 
     Rational:
     ---------
-        CytoSnake requires input from multiple platemaps and barcodes. To ensure 
-        accurate mapping of plate data, it necessitates the use of a barcode as an 
-        input. The barcode files are used by CytoSnake to correctly associate each 
+        CytoSnake requires input from multiple platemaps and barcodes. To ensure
+        accurate mapping of plate data, it necessitates the use of a barcode as an
+        input. The barcode files are used by CytoSnake to correctly associate each
         plate map with its corresponding plate data.
 
     Parameters:
